refactor: move tests to tests dir

+++ src/scripts/tests/test_pdf_to_md.py
@@ -2,9 +2,6 @@ from src import config
 import pymupdf
 import pymupdf4llm
 from pathlib import Path
-from unstructured.partition.pdf import partition_pdf
-from unstructured.chunking.basic import chunk_elements
-from unstructured.chunking.title import chunk_by_title
 from concurrent.futures import ProcessPoolExecutor, as_completed
 import os
 
@@ -22,12 +19,16 @@ def test_nccn():
 
 
 def test_unstructured_nccn(input_pdf_path: Path):
+    from unstructured.partition.pdf import partition_pdf
+    from unstructured.chunking.title import chunk_by_title
+    from unstructured.chunking.basic import chunk_elements
+
     print(f"Processing {input_pdf_path}...")
     elements = partition_pdf(str(input_pdf_path))
     print(f"Found {len(elements)} elements in the PDF.")
 
-    # chunks = chunk_elements(elements, max_characters=1500)
-    chunks = chunk_by_title(elements, max_characters=1500)
+    chunks = chunk_elements(elements, max_characters=1500)
+    # chunks = chunk_by_title(elements, max_characters=1500)
     for i, chunk in enumerate(chunks):
         with open("temp.txt") as f:
             string = f"Chunk {i}:\n\nCategory: {chunk.category}\n\nMetadata: {chunk.metadata.to_dict()}\n\n{chunk.text}"
@@ -39,7 +40,7 @@ def test_unstructured_nccn(input_pdf_path: Path):
         os.remove("temp.txt")  # Remove the temporary file
 
 
-def process_page(args):
+def process_page(args: tuple[Path, int]) -> tuple[int, str]:
     input_pdf_path, page_id = args
     nccn_doc = pymupdf.open(input_pdf_path)
     md_text = pymupdf4llm.to_markdown(nccn_doc, pages=[page_id])
@@ -47,7 +48,9 @@ def process_page(args):
     return page_id, md_text
 
 
-def test_pymupdf_nccn(input_pdf_path: Path, max_workers: int = 8):
+def test_pymupdf_nccn(
+    input_pdf_path: Path, max_workers: int = 8, check: bool = False
+) -> list[str]:
     print(f"Processing {input_pdf_path}...")
     nccn_doc = pymupdf.open(input_pdf_path)
     page_count = nccn_doc.page_count
@@ -63,14 +66,17 @@ def test_pymupdf_nccn(input_pdf_path: Path, max_workers: int = 8):
             print(f"Processed page {page_id + 1} of {page_count}.")
             md_texts[page_id] = md_text
 
-    for i, md_text in enumerate(md_texts):
-        with open("temp.md", "w") as f:
-            f.write(md_text)
-        os.system("bat temp.md")
-        input(f"Press Enter to view page {i + 2}/{len(md_texts)}...")
-        os.system("clear")  # Clear screen before showing next page
-    if Path("temp.md").exists():
-        os.remove("temp.md")  # Remove the temporary file
+    if check:
+        for i, md_text in enumerate(md_texts):
+            with open("temp.md", "w") as f:
+                f.write(md_text)
+            os.system("bat temp.md")
+            input(f"Press Enter to view page {i + 2}/{len(md_texts)}...")
+            os.system("clear")  # Clear screen before showing next page
+        if Path("temp.md").exists():
+            os.remove("temp.md")  # Remove the temporary file
+
+    return md_texts
 
 
 if __name__ == "__main__":
